refactor(app): move the live log pipeline into ainb-app
The Docker log path now lives entirely in ainb-app: the log parser, the
LogEntry model half of live_logs_stream, the JSONL log writer and reader,
the agent-event widgets that turn parsed events into log entries, and
docker::log_streaming, which drives them.

Two terminal ties are cut on the way:
- LogLevel::color returned a ratatui Color. It becomes
  log_parser::level_color in ainb-core, the only place a colour is drawn.
- The widgets' result separator sized itself from crossterm's terminal
  size; it reads ainb_app::viewport instead.

The widgets are string formatters, not draw code, so the render-path
`.await` lint in build.rs stops scanning them. ainb-app picks up
pulldown-cmark and syntect for the widgets.
