+++ ainb-tui/crates/ainb-app/src/components/log_writer.rs
