+++ ainb-tui/crates/ainb-app/src/lib.rs
@@ -32,6 +32,7 @@ pub mod setup;
 pub mod tmux;
 pub mod usage_cache;
 pub mod viewport;
+pub mod widgets;
 
 #[cfg(any(test, feature = "test-support"))]
 pub mod test_support;
