+++ ainb-tui/crates/ainb-app/src/widgets/bash_widget.rs
