+++ ainb-tui/crates/ainb-app/src/widgets/default_widget.rs
