+++ ainb-tui/crates/ainb-app/src/widgets/ls_result_widget.rs
