+++ ainb-tui/crates/ainb-app/src/widgets/mcp_widget.rs
