+++ ainb-tui/crates/ainb-app/src/widgets/read_widget.rs
