+++ ainb-tui/crates/ainb-app/src/widgets/result_parser.rs
