+++ ainb-tui/crates/ainb-app/src/widgets/syntax_highlighter.rs
