+++ ainb-tui/crates/ainb-app/src/widgets/system_reminder_widget.rs
