+++ ainb-tui/crates/ainb-app/src/widgets/thinking_widget.rs
