+++ ainb-tui/crates/ainb-app/src/widgets/tool_result_store.rs
