+++ ainb-tui/crates/ainb-app/src/widgets/unified_message.rs
