+++ ainb-tui/crates/ainb-app/src/widgets/webfetch_widget.rs
