+++ ainb-tui/crates/ainb-app/src/widgets/websearch_widget.rs
