+++ ainb-tui/crates/ainb-app/src/widgets/write_widget.rs
