+++ ainb-tui/crates/ainb-core/src/components/log_formatter.rs
@@ -104,7 +104,7 @@ impl LogFormatter {
         if !matches!(log.level, LogLevel::Info | LogLevel::Debug) {
             spans.push(Span::styled(
                 log.level.icon(),
-                Style::default().fg(log.level.color())
+                Style::default().fg(super::log_parser::level_color(log.level))
             ));
             spans.push(Span::raw(" "));
         }
